不具合関係のAPIの対応
