+++ backend/apps/bug_note/views/defect_view.py
@@ -5,10 +5,12 @@ from drf_spectacular.utils import extend_schema
 
 from apps.bug_note.serializers import DefectCreateSerializer
 from apps.staff_hub.permission import HasUserPermissionObject
-from apps.bug_note.common import check_bug_note_edit_permission
+from apps.bug_note.common import check_bug_note_edit_permission, check_bug_note_view_permission
+from apps.bug_note.models.defect import Defect
+from apps.bug_note.serializers import DefectListSerializer
 
 
-class DefectCreateView(APIView):
+class DefectView(APIView):
     permission_classes = [permissions.IsAuthenticated, HasUserPermissionObject]
 
     @extend_schema(
@@ -25,3 +27,14 @@ class DefectCreateView(APIView):
         defect = serializer.save(create_user=request.user)
 
         return Response(DefectCreateSerializer(defect).data, status=status.HTTP_201_CREATED)
+
+    @extend_schema(
+        responses={200: DefectListSerializer(many=True)},
+        tags=["defect"],
+        description="不具合を一覧で取得"
+    )
+    def get(self, request):
+        check_bug_note_view_permission(request)
+        defects = Defect.get_defects_by_month()
+        serializer = DefectListSerializer(defects, many=True)
+        return Response(serializer.data)
